#49 Update Readme with last modifications and reinclude config tests
